[MPT] Refactoring due to changes in rlp package structure
